+++ src/components/diary/screen/DiaryContainer.jsx
@@ -3,12 +3,9 @@ import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 
 import DiaryCalendar from "../components/DiaryCalendar";
-import EmotionButton from "../components/EmotionButton";
+import DiaryContents from "../components/DiaryContents";
 
-import { ReactComponent as edit } from "../../../img/diary/edit.svg";
-import { ReactComponent as check } from "../../../img/diary/round-check.svg";
 import { __getDiary } from "../../../redux/modules/emotiondiarySlice";
-import { __postDiary } from "../../../redux/modules/emotiondiarySlice";
 
 const DiaryContainer = () => {
   const diaryList = useSelector(
@@ -19,10 +16,10 @@ const DiaryContainer = () => {
   // );
   const dispatch = useDispatch();
 
-  const [diaryContent, setDiaryContent] = useState("");
+  const [diaryContent, setDiaryContent] = useState(diaryList?.diaryContent);
   const [diaryId, setDiaryId] = useState(diaryList?.diaryId);
   const [emotion, setEmotin] = useState(diaryList?.emotion);
-  const [disabled, setDisabled] = useState(true);
+
   const [selectDate, setSelectDate] = useState("");
   const [year, setYear] = useState(new Date().getFullYear());
   const [month, setMonth] = useState(new Date().getMonth() + 1);
@@ -37,19 +34,7 @@ const DiaryContainer = () => {
     emotion: emotion,
   };
 
-  const onEditHandler = () => {
-    setDiaryId(diaryList?.diaryId);
-    setDisabled(false);
-  };
-
-  const onAddHandler = () => {
-    dispatch(__postDiary(DiaryInfo));
-    setDisabled(true);
-  };
-
-  const onDiaryHandler = (e) => {
-    setDiaryContent(e.target.value);
-  };
+  console.log(diaryList);
 
   const onDateHandler = () => {
     const day = ["일", "월", "화", "수", "목", "금", "토"];
@@ -76,42 +61,14 @@ const DiaryContainer = () => {
           setDiaryContent={setDiaryContent}
         />
       </CalendarDiv>
-      <TodoDiv>
-        <DateTitle>{selectDate}</DateTitle>
-        <EmotionDiv>
-          <EmotionBox>
-            <EmotionButton setEmotin={setEmotin} />
-          </EmotionBox>
-          {disabled ? (
-            <EditCheckDiv>
-              <EditCheckBox>
-                <EditBtn onClick={() => onEditHandler()}>
-                  <EditImg />
-                </EditBtn>
-                수정
-              </EditCheckBox>
-            </EditCheckDiv>
-          ) : (
-            <EditCheckDiv>
-              <EditCheckBox>
-                <EditBtn onClick={() => onAddHandler()}>
-                  <CheckImg />
-                </EditBtn>
-                저장
-              </EditCheckBox>
-            </EditCheckDiv>
-          )}
-        </EmotionDiv>
-        {diaryList && (
-          <DiaryText
-            disabled={disabled}
-            value={diaryContent}
-            placeholder="일기를 작성해보세요."
-            onChange={(e) => onDiaryHandler(e)}
-          />
-        )}
-        {diaryList && <DiaryLength>({diaryContent?.length}/200)</DiaryLength>}
-      </TodoDiv>
+      <DiaryContents
+        DiaryInfo={DiaryInfo}
+        diaryContent={diaryContent}
+        setDiaryContent={setDiaryContent}
+        emotion={emotion}
+        setEmotin={setEmotin}
+        selectDate={selectDate}
+      />
     </Container>
   );
 };
@@ -136,101 +93,5 @@ const SelectDiv = styled.div`
   left: 88%;
   align-items: center;
   padding-top: 5px;
-  font-weight: bold;
-`;
-const TodoDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 344px;
-  height: 232px;
-  margin-top: 50px;
-  margin-left: 1px;
-  padding: 15px;
-  border-radius: 16px;
-  background-color: var(--color-default);
-`;
-const DateTitle = styled.span`
-  text-align: center;
-  font-size: 14px;
-  font-weight: bold;
-  margin-top: 2%;
-`;
-
-const EmotionDiv = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 28px;
-  margin-top: 5px;
-`;
-const EmotionBox = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 90%;
-  margin-left: 15px;
-`;
-
-const EditCheckDiv = styled.div`
-  display: flex;
-  width: 10%;
-`;
-const EditCheckBox = styled.div`
-  display: flex;
-  align-items: flex-end;
-  justify-content: flex-end;
-  flex-direction: column;
-  width: 100%;
-  height: 32px;
-  font-size: 10px;
-  font-weight: bold;
-  color: var(--color-gray);
-`;
-
-const EditBtn = styled.button`
-  width: 20px;
-  height: 20px;
-  border: none;
-  background-color: transparent;
-`;
-
-const CheckImg = styled(check)`
-  width: 20px;
-  height: 20px;
-  fill: var(--color-gray);
-`;
-
-const EditImg = styled(edit)`
-  width: 20px;
-  height: 20px;
-`;
-
-const DiaryText = styled.textarea.attrs({ maxLength: 200 })`
-  margin-top: 10px;
-  padding-left: 7px;
-  border: none;
-  background-color: var(--color-default);
-  resize: none;
-  font-size: 12px;
-  font-weight: bold;
-  height: 120px;
-  line-height: 17px;
-  color: var(--color-dark-gray);
-  ::placeholder {
-    text-align: center;
-    font-size: 12px;
-    font-weight: bold;
-  }
-  :focus {
-    outline: none;
-  }
-`;
-
-const DiaryLength = styled.span`
-  display: flex;
-  align-items: center;
-  justify-content: end;
-  color: var(--color-gray);
   font-weight: bold;
 `;
